# Simplify operator classification lookup on Intel