--- a/src/haswell/dom_parser_implementation.cpp
+++ b/src/haswell/dom_parser_implementation.cpp
@@ -70,7 +70,11 @@ simdjson_really_inline json_character_block json_character_block::classify(const
     _mm256_shuffle_epi8(whitespace_table, in.chunks[0]),
     _mm256_shuffle_epi8(whitespace_table, in.chunks[1])
   });
-  const simd8x64<uint8_t> curlified = in.bit_or(0x20); // Turn [ and ] into { and }
+  // Turn [ and ] into { and }
+  const simd8x64<uint8_t> curlified{
+    in.chunks[0] | 0x20,
+    in.chunks[1] | 0x20
+  };
   const uint64_t op = curlified.eq({
     _mm256_shuffle_epi8(op_table, in.chunks[0]),
     _mm256_shuffle_epi8(op_table, in.chunks[1])